chore: Update test:end2end:ldap and test:end2end:server scripts
Update the test:end2end:ldap and test:end2end:server scripts in the package.json file. The changes include exporting the NODE_ENV variable as "end2end" and adding the "--trace-warnings" flag to the node command. This update ensures that the end-to-end server is launched correctly for testing purposes.